[map] Initial support for nogo areas
For now, just displays the polygons. No clickability yet.

+++ lib/main.dart
@@ -166,7 +166,7 @@ class TripPlannerState extends State<TripPlanner> {
   late CancelableOperation<void> getInitialPosition;
   Position? _initialPosition;
 
-  double Function(StationType)? stationPriority;
+  int Function(StationType)? stationPriority;
 
   bool hasModal = false;
